backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose==3.3.0
pydantic==1.10.13
filelock==3.12.2
